+++ examples/data-sources/nxos_ethernet/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_ethernet" "example" {
+}