+++ examples/data-sources/nxos_ethernet_settings/data-source.tf
@@ -1,2 +1,0 @@
-data "nxos_ethernet_settings" "example" {
-}