+++ examples/resources/nxos_ethernet/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ethernet.example "sys/ethpm/inst"