+++ examples/resources/nxos_ethernet/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_ethernet" "example" {
+  mtu = 9216
+}