+++ examples/resources/nxos_ethernet_settings/resource.tf
@@ -1,3 +1,0 @@
-resource "nxos_ethernet_settings" "example" {
-  mtu = 9216
-}