+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Add nxos_subinterface_vrf resource and data source
 - Add nxos_dhcp_relay_interface resource and data source
 - Add nxos_dhcp_relay_address resource and data source
-- Add nxos_ethernet_settings resource and data source
+- Add nxos_ethernet resource and data source
 - Add nxos_ospf resource and data source
 - Add nxos_ospf_instance resource and data source
 - Add nxos_ospf_vrf resource and data source